+++ R/heddle.R
@@ -11,14 +11,29 @@
 #' recycled for every value in your data) or a vector of the same length as
 #' your data (which will be applied element-wise to your data, so that
 #' \code{data[[1]]} will replace \code{pattern[[1]]}).
-#' @param ... If \code{data} is a vector, a string representing the string to
-#' replace with data values in each pattern. If \code{data} is a dataframe, a
-#' set of name = variable pairs, with the name matching the keyword to be
-#' replaced by that variable. Names should be quoted, variable names should
-#' not.
+#' @param ... Values indicating what placeholders in the pattern should be
+#' replaced -- see "Specifying replacement values" below for more.
 #' @param strip.whitespace A boolean (TRUE/FALSE) value indicating if
 #' whitespace should be removed from the replacement variable. Toggle this
 #' on if you're using the variable in chunk labels or similar places.
+#'
+#' @section Replacement values:
+#' \code{heddle} can accept multiple different values for \code{...}, depending
+#' on how you call it.
+#' If \code{data} is a vector (which is the case when either
+#' calling \code{heddle} on a vector directly, or using it in a
+#' \code{\link[dplyr]{mutate}}) call, \code{...} should be unnamed strings
+#' matching the values to be replaced. If any argument passed to \code{...}
+#' isn't found in the pattern, a warning will be raised -- use \code{NA} to
+#' replicate patterns without replacing any values.
+#' If \code{data} is a dataframe (which is the case both when calling
+#' \code{heddle} on a dataframe directly or using it in combination with
+#' \code{\link[tidyr]{nest}} and \code{\link[purrr]{map}}),
+#' \code{...} should be a set of name = variable
+#' pairs, with the name matching the keyword to be replaced by that variable.
+#' Names should be quoted, variable names don't need to be. As with vectors,
+#' if any argument passed to \code{...} isn't found in the pattern, a warning
+#' will be raised.
 #'
 #' @return Returns a character vector of the pattern with placeholders
 #' replaced by variables.
@@ -64,12 +79,29 @@ heddle.default <- function(data, pattern, ..., strip.whitespace = FALSE) {
     data <- gsub("[[:space:]]", "", data)
   }
   if (length(pattern) == 1) {
-    vapply(
-      data,
-      function(x) gsub(dots, x, pattern),
-      FUN.VALUE = character(1)
-    )
+    if (is.na(dots)) {
+      rep(pattern, length(data))
+    } else if (!grepl(dots, pattern)) {
+      warning("Placeholder keyword not found in pattern.")
+    } else {
+      vapply(
+        data,
+        function(x) gsub(dots, x, pattern),
+        FUN.VALUE = character(1)
+      )
+    }
   } else {
+    if (
+      any(
+        vapply(
+          pattern,
+          function(x) grepl(dots, x),
+          logical(1)
+        )
+      ) == FALSE
+    ) {
+      warning("Placeholder keyword not found in pattern.")
+    }
     mapply(function(x, y) gsub(dots, x, y),
       x = data,
       y = pattern,
@@ -96,8 +128,8 @@ heddle.data.frame <- function(data, pattern, ..., strip.whitespace = FALSE) {
       data[, rlang::eval_tidy(dots[[j]], vars)] <- as.character(data[, rlang::eval_tidy(dots[[j]], vars)])
       data[, rlang::eval_tidy(dots[[j]], vars)] <- gsub("[[:space:]]", "", data[, rlang::eval_tidy(dots[[j]], vars)])
     }
-
     for (i in seq_len(nrow(data))) {
+      if (!grepl(names(dots)[[j]], return[[i]])) warning(paste(names(dots)[[j]], "not found in pattern."))
       return[[i]] <- paste(gsub(names(dots)[[j]], data[[i, rlang::eval_tidy(dots[[j]], vars)]], return[[i]]), sep = "", collapse = "")
     }
   }